Yagami.MaterialProperties.refractive_index! — Method

refractive_index!(_, _, _, _, _, _)

Returns the refractive index of air at a given wavelength and temperature, pressure, and humidity. The refractive index is calculated using the specified air model and parameters.

Arguments

model::G: The air model to use for the calculation, which must be a subtype of AirModel.

args::Vararg{AbstractFloat,4}: The parameters for the refractive index calculation, which can be wavelength, temperature, pressure, and humidity.

Returns

The refractive index of air at the specified wavelength and conditions.

Yagami.MaterialProperties.refractive_index — Method

refractive_index(model, params)

Returns the refractive index of air at a given wavelength and temperature, pressure, and humidity. The refractive index is calculated using the specified air model and parameters.

Arguments

model::G: The air model to use for the calculation, which must be a subtype of AirModel.

params::Vararg{AbstractFloat,4}: The parameters for the refractive index calculation, which can be wavelength, temperature, pressure, and humidity.

Returns

The refractive index of air at the specified wavelength and conditions.