Yagami.RayTracing.AtmInterpolate — Type

AtmInteoplate{N,M,T}

A structure representing an atmospheric interpolation object for ray tracing.

AtmInterpolate(knots_θ, knots_h, parameters; logh=false)

Create an atmospheric interpolation object for ray tracing. The interpolation is defined by the knots of height and angle, and the parameters matrix.

Arguments

knots_θ::AbstractVector{T}: A vector of angles in degrees, where T is a subtype of Real.

knots_h::AbstractVector{T}: A vector of heights in kilometers, where T is a subtype of Real.

parameters::AbstractMatrix{T}: A matrix of parameters for the interpolation, where T is a subtype of Real.

logh::Bool: If true, use logarithmic interpolation for height. Default is false.

Returns

An AtmInterpolate object with the specified knots and parameters.

Yagami.RayTracing.AtmosphereSetting — Type

AtmosphereSetting(knots_θ, knots_h, temperature, pressure, humidity=0.0, co2ppm=0.0, wavelength=10.0)

Create an atmosphere setting interpolation object for ray tracing.

Arguments

knots_θ::AbstractVector{T}: A vector of angles in degrees, where T is a subtype of Real.

knots_h::AbstractVector{T}: A vector of heights in kilometers, where T is a subtype of Real.

temperature::AbstractMatrix{T}: A matrix of temperature values, where each row corresponds to a level and each column corresponds to a radius.

pressure::AbstractMatrix{T}: A matrix of pressure values, where each row corresponds to a level and each column corresponds to a radius.

humidity::Real (optional): A scalar or matrix of humidity values, default is 0.0. If a scalar is provided, it is broadcasted to match the size of the temperature matrix.

co2ppm::Real (optional): A scalar or matrix of CO2 concentration values in parts per million, default is 0.0. If a scalar is provided, it is broadcasted to match the size of the temperature matrix.

wavelength::Real (optional): A scalar or matrix of wavelength values, default is 10.0. If a scalar is provided, it is broadcasted to match the size of the temperature matrix.

Returns

An AtmosphereSetting object containing interpolated atmospheric properties for ray tracing.

Yagami.RayTracing.Ray2D — Type

`Ray2D{T}`

Data structure for ray tracing in 2D space.

Arguments

pointx::T: x-coordinate of the ray's starting point in km. Default is 0.0.

pointy::T: y-coordinate of the ray's starting point in km. Default is 0.0.

directionx::T: x-component of the ray's normalized direction vector. Default is 0.0.

directiony::T: y-component of the ray's normalized direction vector. Default is 1.0.

i::Int: index of the wedge in which the ray is currently located. Default is 0.

j::Int: index of the sub-wedge in which the ray is currently located. Default is 0.

Yagami.RayTracing.Ray2D — Method

Ray2D()
Ray2D(pointx)
Ray2D(pointx, pointy)
Ray2D(pointx, pointy, directionx)
Ray2D(pointx, pointy, directionx, directiony)
Ray2D(pointx, pointy, directionx, directiony, i)
Ray2D(pointx, pointy, directionx, directiony, i, j)

Data structures for ray tracing in 2D. Ray2D represents a ray in 2D space with a point and direction and the current wedge index

Arguments

pointx::T: x-coordinate of the ray's starting point in km. Default is 0.0.

pointy::T: y-coordinate of the ray's starting point in km. Default is 0.0.

directionx::T: x-component of the ray's normalized direction vector. Default is 0.0.

directiony::T: y-component of the ray's normalized direction vector. Default is 1.0.

i::Int: index of the wedge in which the ray is currently located. Default is 0.

j::Int: index of the sub-wedge in which the ray is currently located. Default is 0.

Yagami.RayTracing.ResultRay — Type

ResultRay{T}

Data structure to hold the result of a ray intersection in 2D space. This structure contains the intersection point, direction, index of the wedge, length of the intersection, altitude, and azimuth of the ray.

Arguments

pointx::A: x-coordinate of the intersection point in km.

pointy::A: y-coordinate of the intersection point in km.

directionx::A: x-component of the intersection direction vector.

directiony::A: y-component of the intersection direction vector.

index_i::Ai: index of the wedge in which the ray intersects.

index_j::Ai: index of the sub-wedge in which the ray intersects.

length::A: length of the intersection in km.

altitude::A: altitude of the intersection in km.

azimuth::A: azimuth angle of the intersection in degrees.

Yagami.RayTracing.__setaltitude! — Method

__setaltitude!(r::ResultRay, value) Set the :altitude field of the ResultRay.

Yagami.RayTracing.__setazimuth! — Method

__setazimuth!(r::ResultRay, value) Set the :azimuth field of the ResultRay.

Yagami.RayTracing.__setdescending! — Method

__setdescending!(df::DistanceFunc, value::Bool=true) Set the :descending field of the DistanceFunc.

Yagami.RayTracing.__setdirectionx! — Method

__setdirectionx!(r::ResultRay, value) Set the :directionx field of the ResultRay.

Yagami.RayTracing.__setdirectionx! — Method

__setdirectionx!(df::DistanceFunc, value::T=0.0) Set the :directionx field of the DistanceFunc.

Yagami.RayTracing.__setdirectiony! — Method

__setdirectiony!(r::ResultRay, value) Set the :directiony field of the ResultRay.

Yagami.RayTracing.__setdirectiony! — Method

__setdirectiony!(df::DistanceFunc, value::T=1.0) Set the :directiony field of the DistanceFunc.

Yagami.RayTracing.__sethmax! — Method

__sethmax!(df::DistanceFunc, value::T=120.0) Set the :hmax field of the DistanceFunc.

Yagami.RayTracing.__sethmin! — Method

__sethmin!(df::DistanceFunc, value::T=0.0) Set the :hmin field of the DistanceFunc.

Yagami.RayTracing.__seti! — Method

__seti!(df::DistanceFunc, value::Int=0) Set the :i field of the DistanceFunc.

Yagami.RayTracing.__setindex_i! — Method

__setindex_i!(r::ResultRay, value) Set the :index_i field of the ResultRay.

Yagami.RayTracing.__setindex_j! — Method

__setindex_j!(r::ResultRay, value) Set the :index_j field of the ResultRay.

Yagami.RayTracing.__setislevel! — Method

__setislevel!(df::DistanceFunc, value::Bool=true) Set the :islevel field of the DistanceFunc.

Yagami.RayTracing.__setj! — Method

__setj!(df::DistanceFunc, value::Int=0) Set the :j field of the DistanceFunc.

Yagami.RayTracing.__setlength! — Method

__setlength!(r::ResultRay, value) Set the :length field of the ResultRay.

Yagami.RayTracing.__setpointx! — Method

__setpointx!(r::ResultRay, value) Set the :pointx field of the ResultRay.

Yagami.RayTracing.__setpointx! — Method

__setpointx!(df::DistanceFunc, value::T=1.0e-5) Set the :pointx field of the DistanceFunc.

Yagami.RayTracing.__setpointy! — Method

__setpointy!(r::ResultRay, value) Set the :pointy field of the ResultRay.

Yagami.RayTracing.__setpointy! — Method

__setpointy!(df::DistanceFunc, value::T=1.0e-5) Set the :pointy field of the DistanceFunc.

Yagami.RayTracing.__setθmax! — Method

__setθmax!(df::DistanceFunc, value::T=Inf) Set the :θmax field of the DistanceFunc.

Yagami.RayTracing.__setθmin! — Method

__setθmin!(df::DistanceFunc, value::T=-Inf) Set the :θmin field of the DistanceFunc.

Yagami.RayTracing.create_atmosphere — Method

create_atmosphere(; θᵢ, hᵢ, temperatureᵢ, pressureᵢ, humidity=0.0, co2ppm=0.0, wavelength=10.0, knots_θ=create_hlevelset(T;), knots_h=create_radii(T;))

Create an atmosphere setting for ray tracing. This function generates an AtmosphereSetting object based on the provided parameters, which include angles, heights, temperature, pressure, humidity, CO2 concentration, and wavelength.

Key Arguments

θᵢ::AbstractVector{T}: A vector of angles in degrees, where T is a subtype of Real.

hᵢ::AbstractVector{T}: A vector of heights in kilometers, where T is a subtype of Real.

temperatureᵢ::AbstractMatrix{T}: A matrix of temperature values, where each row corresponds to a level and each column corresponds to a radius.

pressureᵢ::AbstractMatrix{T}: A matrix of pressure values, where each row corresponds to a level and each column corresponds to a radius.

humidity=0.0: A scalar or matrix of humidity values, default is 0.0. If a scalar is provided, it is broadcasted to match the size of the temperature matrix.

co2ppm=0.0: A scalar or matrix of CO2 concentration values in parts per million, default is 0.0. If a scalar is provided, it is broadcasted to match the size of the temperature matrix.

wavelength=10.0: A scalar or matrix of wavelength values, default is 10.0. If a scalar is provided, it is broadcasted to match the size of the temperature matrix.

knots_θ::AbstractVector{T}=create_hlevelset(T;): A vector of angles in degrees for the knots, default is created using create_hlevelset.

knots_h::AbstractVector{T}=create_radii(T;): A vector of heights in kilometers for the knots, default is created using create_radii.

Returns

An AtmosphereSetting object containing interpolated atmospheric properties for ray tracing.

Yagami.RayTracing.create_hlevelset — Method

create_hlevelset(; ...)
create_hlevelset(; hmin, hmax, levels, logscale)

Create a set of levels for the raytracing method. The levels are evenly spaced in the range [hmin, hmax] if logscale=false, or logarithmically spaced if logscale=true.

Note

The levels are sorted in descending order, meaning the first element is the maximum level and the last element is the minimum level. This is the format required by the reay tracing algorithm.

Arguments

T::Type{<:Real}: The type of the levels, default is Float64.

hmin::T: The minimum value of the levels, default is 0.0.

hmax::T: The maximum value of the levels, default is 120.0.

levels::Int: The number of levels to create, must be a positive integer, default is 100.

logscale::Bool: If true, the levels are logarithmically spaced; if false, they are linearly spaced. Default is true.

Returns

A vector of levels of type T, either logarithmically or linearly spaced between hmin and hmax.

Yagami.RayTracing.create_radii — Method

create_radii(; ...)
create_radii(; θmin, θmax, radii)

Create a set of radii for the raytracing method. The radii are evenly spaced in the range [θmin, θmax].

Note

The radii are sorted in ascending order, meaning the first element is the minimum radius and the last element is the maximum radius.

Arguments

T::Type{<:Real}: The type of the radii, default is Float64.

θmin::Real: The minimum angle in degrees, default is 0.0.

θmax::Real: The maximum angle in degrees, default is 359.0.

radii::Int: The number of radii to create, must be a positive integer, default is 360.

Returns

A vector of radii of type T, evenly spaced between θmin and θmax.

Yagami.RayTracing.extrapolateco2ppm — Method

`extrapolateco2ppm(prob::RayTracingProblem, θ::Real, h::Real)`
`extrapolateco2ppm(prob::RayTracingProblem; theta=0.0, h=0.0)`

Extrapolates the CO2 concentration in ppm at a given line of angle θ and height h for a given RayTracingProblem.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

θ::Real: The angle in degree and geodesic coordinates.

h::Real: The altitude with respect to the surface in km.

Returns:

Float64: The extrapolated CO2 concentration in ppm

Key Arguments:

theta::Real: The angle in degrees (default is 0.0).

h::Real: The height in kilometers (default is 0.0).

Example:

julia> extrapolateco2ppm(prob, 10.0, 5.0) # angle in degrees, height in km
400.0
julia> extrapolateco2ppm(prob; theta=10.0, h=5.0) # using keyword arguments
400.0
julia> extrapolateco2ppm(prob) # defaults to theta=0.0, h=0.0
400.0

Yagami.RayTracing.extrapolatehumidity — Method

`extrapolatehumidity(prob::RayTracingProblem, θ::Real, h::Real)`
`extrapolatehumidity(prob::RayTracingProblem; theta=0.0, h=0.0)`

Extrapolates the humidity at a given line of angle θ and height h for a given RayTracingProblem.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

θ::Real: The angle in degree and geodesic coordinates.

h::Real: The altitude with respect to the surface in km.

Returns:

Float64: The extrapolated humidity in %

Yagami.RayTracing.extrapolatepressure — Method

extrapolatepressure(prob::RayTracingProblem, θ::Real, h::Real) extrapolatepressure(prob::RayTracingProblem; theta=0.0, h=0.0) Extrapolates the pressure at a given line of angle θ and height h for a given RayTracingProblem.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

θ::Real: The angle in degree and geodesic coordinates.

h::Real: The altitude with respect to the surface in km.

Returns:

Float64: The extrapolated pressure in Pa

Key Arguments:

theta::Real: The angle in degrees (default is 0.0).

h::Real: The height in kilometers (default is 0.0).

Example:

julia
> extrapolatepressure(prob, 10.0, 5.0) # angle in degrees, height in km
101325.0
julia> extrapolatepressure(prob; theta=10.0, h=5.0) # using keyword arguments
101325.0
julia> extrapolatepressure(prob) # defaults to theta=0.0, h=0.0
101325.0

Yagami.RayTracing.extrapolatetemperature — Method

extrapolatetemperature(prob::RayTracingProblem, θ::Real, h::Real) extrapolatetemperature(prob::RayTracingProblem; theta=0.0, h=0.0) Extrapolates the temperature at a given line of angle θ and height h for a given RayTracingProblem.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

θ::Real: The angle in degree and geodesic coordinates.

h::Real: The altitude with respect to the surface in km.

Returns:

Float64: The extrapolated temperature in Kelvin

Key Arguments:

theta::Real: The angle in degrees (default is 0.0).

h::Real: The height in kilometers (default is 0.0).

Example:

julia> extrapolatetemperature(prob, 10.0, 5.0) # angle in degrees, height in km
273.15
julia> extrapolatetemperature(prob; theta=10.0, h=5.0) # using keyword arguments
275.15
julia> extrapolatetemperature(prob) # defaults to theta=0.0, h=0.0
278.15

Yagami.RayTracing.extrapolatewavelength — Method

`extrapolatewavelength(prob::RayTracingProblem, θ::Real, h::Real)`
`extrapolatewavelength(prob::RayTracingProblem; theta=0.0, h=0.0)`

Extrapolates the wavelength at a given line of angle θ and height h for a given RayTracingProblem.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

θ::Real: The angle in degree and geodesic coordinates.

h::Real: The altitude with respect to the surface in km.

Returns:

Float64: The extrapolated wavelength in μm

Key Arguments:

theta::Real: The angle in degrees (default is 0.0).

h::Real: The height in kilometers (default is 0.0).

Example:

julia> extrapolatewavelength(prob, 10.0, 5.0) # angle in degrees, height in km
1.0
julia> extrapolatewavelength(prob; theta=10.0, h=5.0) # using keyword arguments
1.0
julia> extrapolatewavelength(prob) # defaults to theta=0.0, h=0.0
1.0

Yagami.RayTracing.getaltitude — Method

getaltitude(r::ResultRay, int::Int) Return the :altitude field of the ResultRay.

Yagami.RayTracing.getazimuth — Method

getazimuth(r::ResultRay, int::Int) Return the :azimuth field of the ResultRay.

Yagami.RayTracing.getdescending — Method

getdescending(df::DistanceFunc) Return the :descending field of the DistanceFunc.

Yagami.RayTracing.getdirection — Method

getdirection(r)
getdirection(df)

Get the direction of a Ray2D as a 2D vector. Returns a SVector{2,T} where T is the type of the ray's direction.

Yagami.RayTracing.getdirection — Method

getdirection(r, ind)
getdirection(df)

Get the x or y component of the direction of a Ray2D.

ind::Int: Either 1 for x-component or 2 for y-component.

Yagami.RayTracing.getdirection — Method

getdirection(r, sym)
getdirection(df)

Get the x or y component of the direction of a Ray2D.

sym::Symbol: Either :x or :y to specify which component to return.

Yagami.RayTracing.getdirectionx — Method

getdirectionx(df::DistanceFunc) Return the :directionx field of the DistanceFunc.

Yagami.RayTracing.getdirectionx — Method

getdirectionx(r::ResultRay, int::Int) Return the :directionx field of the ResultRay.

Yagami.RayTracing.getdirectiony — Method

getdirectiony(df::DistanceFunc) Return the :directiony field of the DistanceFunc.

Yagami.RayTracing.getdirectiony — Method

getdirectiony(r::ResultRay, int::Int) Return the :directiony field of the ResultRay.

Yagami.RayTracing.gethmax — Method

gethmax(df::DistanceFunc) Return the :hmax field of the DistanceFunc.

Yagami.RayTracing.gethmin — Method

gethmin(df::DistanceFunc) Return the :hmin field of the DistanceFunc.

Yagami.RayTracing.geti — Method

geti(df::DistanceFunc) Return the :i field of the DistanceFunc.

Yagami.RayTracing.getindex_i — Method

getindex_i(r::ResultRay, int::Int) Return the :index_i field of the ResultRay.

Yagami.RayTracing.getindex_j — Method

getindex_j(r::ResultRay, int::Int) Return the :index_j field of the ResultRay.

Yagami.RayTracing.getislevel — Method

getislevel(df::DistanceFunc) Return the :islevel field of the DistanceFunc.

Yagami.RayTracing.getj — Method

getj(df::DistanceFunc) Return the :j field of the DistanceFunc.

Yagami.RayTracing.getlength — Method

getlength(r::ResultRay, int::Int) Return the :length field of the ResultRay.

Yagami.RayTracing.getpoint — Method

getpoint(r)
getpoint(df)

Get the point and direction of a Ray2D as a 2D vector. Returns a SVector{2,T} where T is the type of the ray's coordinates.

Yagami.RayTracing.getpoint — Method

getpoint(r, ind)
getpoint(df)

Get the x or y coordinate of the point of a Ray2D.

ind::Int: Either 1 for x-coordinate or 2 for y-coordinate.

Yagami.RayTracing.getpoint — Method

getpoint(r, sym)
getpoint(df)

Get the x or y coordinate of the point of a Ray2D.

sym::Symbol: Either :x or :y to specify which coordinate to return.

Yagami.RayTracing.getpointx — Method

getpointx(df::DistanceFunc) Return the :pointx field of the DistanceFunc.

Yagami.RayTracing.getpointx — Method

getpointx(r::ResultRay, int::Int) Return the :pointx field of the ResultRay.

Yagami.RayTracing.getpointy — Method

getpointy(df::DistanceFunc) Return the :pointy field of the DistanceFunc.

Yagami.RayTracing.getpointy — Method

getpointy(r::ResultRay, int::Int) Return the :pointy field of the ResultRay.

Yagami.RayTracing.getquotetangent — Method

getquotetangent(prob::RayTracingProblem, los::Int, scan::Int) getquotetangent(prob::RayTracingProblem; los=1, scan=1) Returns the tangent point coordinates for a given line of sight (los) and scan index in a RayTracingProblem.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

los::Int: The line of sight index.

scan::Int: The scan index.

Returns:

Tuple{Float64, Float64}: A tuple containing the tangent height and azimuth angle at the specified los and scan.

Key Arguments:

los::Int: The line of sight index (default is 1).

scan::Int: The scan index (default is 1).

Example:

julia> getquotetangent(prob, 1, 1) # returns the tangent point for los=1 and scan=1
(100.0, 45.0)
julia> getquotetangent(prob; los=1, scan=1) # using keyword arguments
(100.0, 45.0)
julia> getquotetangent(prob) # defaults to los=1, scan=1
(100.0, 45.0)

Yagami.RayTracing.getsatdirection — Method

getsatdirection(prob::RayTracingProblem, los::Int, scan::Int) getsatdirection(prob::RayTracingProblem; los=1, scan=1) Returns the satellite direction (from the nadir angle) for a given line of sight (los) and scan index in a RayTracingProblem.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

los::Int: The line of sight index.

scan::Int: The scan index.

Returns:

Vector{Float64}: A vector containing the x and y components of the satellite direction.

Key Arguments:

los::Int: The line of sight index (default is 1).

scan::Int: The scan index (default is 1).

Example:

julia julia> getsatdirection(prob, 1, 1) # returns the direction for los=1 and scan=1 [0.1, 0.2] julia> getsatdirection(prob; los=1, scan=1) # using keyword arguments [0.1, 0.2] julia> getsatdirection(prob) # defaults to los=1, scan=1 [0.1, 0.2]`

Yagami.RayTracing.getsatposition — Method

getsatposition(prob::RayTracingProblem, los::Int, scan::Int) getsatposition(prob::RayTracingProblem; los=1, scan=1) Returns the satellite initial position for a given line of sight (los) and scan index in a RayTracingProblem.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

los::Int: The line of sight index.

scan::Int: The scan index.

Returns:

Vector{Float64}: A vector containing the x and y coordinates of the satellite position.

Key Arguments:

los::Int: The line of sight index (default is 1).

scan::Int: The scan index (default is 1).

Example:

```julia julia

getsatposition(prob, 1, 1) # returns the position for los=1 and scan=1

[100.0, 200.0] julia> getsatposition(prob; los=1, scan=1) # using keyword arguments [100.0, 200.0] julia> getsatposition(prob) # defaults to los=1, scan=1 [100.0, 200.0]

Yagami.RayTracing.getwedgeindex — Method

getwedgeindex(r, ind)

Get the i or j index of the wedge of a Ray2D.

ind::Int: Either 1 for i-index or 2 for j-index.

Yagami.RayTracing.getwedgeindex — Method

getwedgeindex(r, sym)

Get the i or j index of the wedge of a Ray2D.

sym::Symbol: Either :i or :j to specify which index to return.

Yagami.RayTracing.getθmax — Method

getθmax(df::DistanceFunc) Return the :θmax field of the DistanceFunc.

Yagami.RayTracing.getθmin — Method

getθmin(df::DistanceFunc) Return the :θmin field of the DistanceFunc.

Yagami.RayTracing.node_co2ppm — Method

node_co2ppm(prob::RayTracingProblem, i_theta::Int, j_h::Int) Returns the nodes of the co2ppm for a given RayTracingProblem at specified wedge edge.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

i_theta::Int: The index of the theta knot (angle).

j_h::Int: The index of the h knot (height).

Returns:

Tuple{Int,Int},T: A tuple containing the theta and h values at the specified indices and the value of co2ppm.

Example:

julia> node_co2ppm(prob, 1, 1) # returns the nodes for i_theta=1 and j_h=1
(0.0, 0.0, 273.15)

Yagami.RayTracing.node_humidity — Method

node_humidity(prob::RayTracingProblem, i_theta::Int, j_h::Int) Returns the nodes of the humidity for a given RayTracingProblem at specified wedge edge.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

i_theta::Int: The index of the theta knot (angle).

j_h::Int: The index of the h knot (height).

Returns:

Tuple{Int,Int},T: A tuple containing the theta and h values at the specified indices and the value of humidity.

Example:

julia> node_humidity(prob, 1, 1) # returns the nodes for i_theta=1 and j_h=1
(0.0, 0.0, 273.15)

Yagami.RayTracing.node_pressure — Method

node_pressure(prob::RayTracingProblem, i_theta::Int, j_h::Int) Returns the nodes of the pressure for a given RayTracingProblem at specified wedge edge.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

i_theta::Int: The index of the theta knot (angle).

j_h::Int: The index of the h knot (height).

Returns:

Tuple{Int,Int},T: A tuple containing the theta and h values at the specified indices and the value of pressure.

Example:

julia> node_pressure(prob, 1, 1) # returns the nodes for i_theta=1 and j_h=1
(0.0, 0.0, 273.15)

Yagami.RayTracing.node_temperature — Method

node_temperature(prob::RayTracingProblem, i_theta::Int, j_h::Int) Returns the nodes of the temperature for a given RayTracingProblem at specified wedge edge.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

i_theta::Int: The index of the theta knot (angle).

j_h::Int: The index of the h knot (height).

Returns:

Tuple{Int,Int},T: A tuple containing the theta and h values at the specified indices and the value of temperature.

Example:

julia> node_temperature(prob, 1, 1) # returns the nodes for i_theta=1 and j_h=1
(0.0, 0.0, 273.15)

Yagami.RayTracing.node_wavelength — Method

node_wavelength(prob::RayTracingProblem, i_theta::Int, j_h::Int) Returns the nodes of the wavelength for a given RayTracingProblem at specified wedge edge.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

i_theta::Int: The index of the theta knot (angle).

j_h::Int: The index of the h knot (height).

Returns:

Tuple{Int,Int},T: A tuple containing the theta and h values at the specified indices and the value of wavelength.

Example:

julia> node_wavelength(prob, 1, 1) # returns the nodes for i_theta=1 and j_h=1
(0.0, 0.0, 273.15)

Yagami.RayTracing.num_edges — Method

num_edges(prob::RayTracingProblem) Returns the number of edges in a RayTracingProblem as N x M, where N is the number of angle knots and M is the number of height knots.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

Returns:

Tuple{Int, Int}: A tuple containing the number of edges (N, M).

Note:

Due to the circularity of the matrix along the theta dimension, the number of edges in the theta direction is one less than the number of knots in the theta direction, while the number of edges in the height dimension is one less than the number of knots in the height direction.

Example:

julia> num_edges(prob) # returns the number of edges in the problem
(10, 19)

See also: num_wedges

Yagami.RayTracing.num_wedges — Method

numwedges(prob::RayTracingProblem) Returns the number of wedges in a RayTracingProblem as N x M, where N is the number of angle knots and M is the number of height knots.

Arguments:

prob::RayTracingProblem: The ray tracing problem instance.

Returns:

Tuple{Int, Int}: A tuple containing the number of wedges (N, M).

Note:

due to the circularity of the matrix along the theta dimension, the number of wedges in the theta dimension is the same as the number of edges in the theta direction, while the number of wedges in the height dimension is one less than the number of edges in the h direction.

Example:

julia> numwedges(prob) # returns the number of wedges in the problem
(10, 20)

See also: num_edges

Yagami.RayTracing.wedge_refractive — Method

wedge_refractive(problem::RayTracingProblem, i_theta::Int, j_h::Int;complement::Bool=false) Returns the refractive index at a given wedge edge specified by indices i_theta and j_h in a RayTracingProblem.

Arguments:

problem::RayTracingProblem: The ray tracing problem instance.

i_theta::Int: The index of the theta knot (angle).

j_h::Int: The index of the h knot (height).

Key Arguments:

complement::Bool: If true, returns the complement of the refractive index (n-1), that can be more useful in some cases (default is false).

Returns:

Tuple{Tuple{Float64, Float64}, Float64}: A tuple containing the theta and h values at the specified indices and the refractive index.

Example:

julia> wedge_refractive(prob, 1, 1) # returns the refractive index for i_theta=1 and j_h=1
((0.0, 0.0), 1.0003)